refactor(app): update database config files
